requirements/test.txt
coverage[toml]==7.6.0
factory-boy==3.3.0
requests_mock==1.12.1
beautifulsoup4==4.12.3
pytest==8.3.1
pytest-django==4.8.0
pytest-xdist==3.6.1
pytest-cov==5.0.0
playwright==1.45.0
pytest-playwright==0.5.1
pytest-subtests==0.13.1
